fix(ci): python-full-coverage — Cursor wiring left executor compatibility, docs, and line guards out of sync
