+++ src/awf/service/provider_readiness.py
@@ -482,168 +482,6 @@ def _agent_runtime_cli_executable(provider: ProviderName) -> str | None:
     }.get(provider)
 
 
-def _agent_runtime_cli_reason_prefix(provider: ProviderName) -> str:
-    return {
-        "codex": "CODEX",
-        "claude_code": "CLAUDE",
-        "cursor": "CURSOR",
-        "gemini": "GEMINI",
-        "opencode": "OPENCODE",
-        "grok": "GROK",
-    }.get(provider, "PROVIDER")
-
-
-def _probe_agent_runtime_cli(
-    settings: ServiceSettings,
-    *,
-    executable: str,
-    provider: ProviderName,
-    environ: Mapping[str, str],
-    run_subprocess: SubprocessRun,
-    secrets: frozenset[str],
-) -> dict[str, Any]:
-    reason_prefix = _agent_runtime_cli_reason_prefix(provider)
-    args = [
-        "docker",
-        "run",
-        "--rm",
-        "--entrypoint",
-        "sh",
-        settings.agent_runtime_image,
-        "-lc",
-        f"command -v {executable}",
-    ]
-    try:
-        result = run_subprocess(
-            args,
-            check=False,
-            capture_output=True,
-            text=True,
-            timeout=_PROVIDER_PROBE_TIMEOUT_SECONDS,
-            env=environ,
-        )
-    except FileNotFoundError:
-        return {
-            "status": "fail",
-            "reason_code": "DOCKER_CLI_NOT_FOUND",
-            "message": (
-                "Docker CLI was not found while probing the configured agent runtime image."
-            ),
-        }
-    except subprocess.TimeoutExpired:
-        return {
-            "status": "fail",
-            "reason_code": f"{reason_prefix}_RUNTIME_CLI_PROBE_TIMEOUT",
-            "message": (
-                f"Agent runtime CLI probe for {executable!r} exceeded "
-                f"{_PROVIDER_PROBE_TIMEOUT_SECONDS:g}s."
-            ),
-        }
-    except Exception as exc:
-        _log_redacted_exception(
-            "provider_readiness.agent_runtime_cli_probe_exception",
-            exc,
-            secrets,
-        )
-        detail = _redact(_truncate(f"{type(exc).__name__}: {exc}"), secrets)
-        return {
-            "status": "fail",
-            "reason_code": f"{reason_prefix}_RUNTIME_CLI_PROBE_ERROR",
-            "message": "Agent runtime CLI probe failed before completion.",
-            "detail": detail,
-        }
-
-    if result.returncode == 0:
-        return {
-            "status": "ok",
-            "reason_code": f"{reason_prefix}_RUNTIME_CLI_AVAILABLE",
-            "detail": _redact(_truncate(result.stdout.strip()), secrets)
-            if result.stdout.strip()
-            else None,
-        }
-
-    detail = _redact(
-        _truncate(result.stderr or result.stdout or f"{executable} was not found"),
-        secrets,
-    )
-    return {
-        "status": "fail",
-        "reason_code": f"{reason_prefix}_RUNTIME_CLI_NOT_FOUND",
-        "message": (
-            f"The configured agent runtime image {settings.agent_runtime_image!r} "
-            f"does not expose the {executable!r} CLI required by provider {provider!r}."
-        ),
-        "detail": detail,
-    }
-
-
-def _probe_cli_auth_status(
-    *,
-    provider_label: str,
-    args: list[str],
-    failure_reason: str,
-    timeout_reason: str,
-    missing_reason: str,
-    error_reason: str,
-    environ: Mapping[str, str],
-    run_subprocess: SubprocessRun,
-    secrets: frozenset[str],
-) -> dict[str, Any]:
-    try:
-        result = run_subprocess(
-            args,
-            check=False,
-            capture_output=True,
-            text=True,
-            timeout=_PROVIDER_PROBE_TIMEOUT_SECONDS,
-            env=environ,
-        )
-    except FileNotFoundError:
-        return {
-            "status": "fail",
-            "reason_code": missing_reason,
-            "message": f"{provider_label} CLI was not found for auth status probing.",
-        }
-    except subprocess.TimeoutExpired:
-        return {
-            "status": "fail",
-            "reason_code": timeout_reason,
-            "message": (
-                f"{provider_label} auth status probe exceeded {_PROVIDER_PROBE_TIMEOUT_SECONDS:g}s."
-            ),
-        }
-    except Exception as exc:
-        _log_redacted_exception(
-            "provider_readiness.cli_auth_probe_exception",
-            exc,
-            secrets,
-        )
-        detail = _redact(_truncate(f"{type(exc).__name__}: {exc}"), secrets)
-        return {
-            "status": "fail",
-            "reason_code": error_reason,
-            "message": f"{provider_label} auth status probe failed before completion.",
-            "detail": detail,
-        }
-
-    if result.returncode == 0:
-        return {
-            "status": "ok",
-            "reason_code": f"{failure_reason.removesuffix('_FAILED')}_OK",
-        }
-
-    detail = _redact(
-        _truncate(result.stderr or result.stdout or "auth status exited non-zero"),
-        secrets,
-    )
-    return {
-        "status": "fail",
-        "reason_code": failure_reason,
-        "message": f"{provider_label} auth status probe reported unusable auth.",
-        "detail": detail,
-    }
-
-
 def _preflight_reason_code(
     *,
     provider_result: Mapping[str, Any],
@@ -1192,15 +1030,6 @@ def _check_cursor_readiness(
     return result
 
 
-def _runtime_cli_probe_payload(probe: Mapping[str, Any]) -> dict[str, str]:
-    """Return the public string fields from a runtime CLI probe result."""
-    return {
-        key: value
-        for key in ("status", "reason_code", "message", "detail")
-        if isinstance((value := probe.get(key)), str) and value
-    }
-
-
 def _check_gemini(
     *,
     environ: Mapping[str, str],
@@ -1608,12 +1437,15 @@ from awf.service.provider_readiness_helpers import (  # noqa: E402
     _ordered_names,
     _primary_credential_scope,
     _primary_isolation,
+    _probe_agent_runtime_cli,
+    _probe_cli_auth_status,
     _probe_ollama,
     _probe_ollama_model,
     _redact,
     _redact_with_redaction_parts,
     _redacted_warning,
     _run_subprocess,
+    _runtime_cli_probe_payload,
     _secret_values,
     _security_summary,
     _security_warning,
@@ -1630,5 +1462,6 @@ __all__ = [
     "selected_provider_readiness_preflight",
     "validate_provider_names",
     "_redact",
+    "_probe_cli_auth_status",
     "_secret_values",
 ]
