+++ tests/unit/control/test_executor_monitor_recovery_parts/test_executor_monitor_recovery_part_002.py
@@ -1468,48 +1468,3 @@ async def test_rebase_only_recovery_marks_operation_failed_when_recording_raises
     assert rebase_ops[0].error_message == "write exploded"
     assert isinstance(rebase_ops[0].result, dict)
     assert rebase_ops[0].result["reason_code"] == "MONITOR_RECOVERY_REBASE_FAILED"
-
-
-@pytest.mark.unit
-async def test_operator_rebase_operation_is_reused_and_failed_when_rebase_fails(
-    fake: FakeCommandRunner,
-    factory: async_sessionmaker[AsyncSession],
-    tmp_path: Path,
-) -> None:
-    executor = _make_executor(fake=fake, factory=factory, tmp_path=tmp_path)
-    ws_id = await _seed_ready_workspace_with_recovery(
-        factory,
-        recovery_mode="rebase_only",
-        source="operator_api",
-        operation_type=OperationType.rebase,
-    )
-
-    fake.queue_result(returncode=0)  # git fetch origin <base>
-    fake.queue_result(returncode=0)  # git switch <branch>
-    fake.queue_result(returncode=1)  # git merge-base --is-ancestor origin/<base> HEAD
-    fake.queue_result(returncode=1, stderr="conflict on README.md")  # git rebase
-    fake.queue_result(returncode=0)  # git rebase --abort
-
-    await executor.execute(ws_id)
-
-    assert _all_adapter_args(fake) == []
-    async with factory() as s:
-        ws = await WorkspaceRepository(s).get(ws_id)
-        ops = await OperationRepository(s).list_all(workspace_id=ws_id)
-    assert ws is not None
-    assert ws.status == WorkspaceStatus.failed.value
-    rebase_ops = [op for op in ops if op.type == OperationType.rebase.value]
-    assert len(rebase_ops) == 1
-    assert rebase_ops[0].payload is not None
-    assert rebase_ops[0].payload["source"] == "operator_api"
-    assert rebase_ops[0].status == OperationStatus.failed.value
-    assert rebase_ops[0].started_at is not None
-    assert rebase_ops[0].finished_at is not None
-    assert rebase_ops[0].error_code == "MONITOR_RECOVERY_REBASE_FAILED"
-    assert "conflict on README.md" in (rebase_ops[0].error_message or "")
-    assert rebase_ops[0].result == {
-        "status": "failed",
-        "reason_code": "MONITOR_RECOVERY_REBASE_FAILED",
-        "source_base_sha": "a" * 40,
-        "source_head_sha": "d" * 40,
-    }
