+++ tests/unit/service/test_provider_readiness_parts/test_provider_readiness_part_001.py
@@ -5,7 +5,6 @@ from __future__ import annotations
 import json
 import logging
 import subprocess
-from dataclasses import replace
 from pathlib import Path
 from types import SimpleNamespace
 from typing import Any
@@ -1432,258 +1431,3 @@ def test_provider_readiness_codex_missing_warns_by_default_and_fails_when_strict
     assert strict_payload["strict_providers"] == ["codex"]
     assert strict_codex["status"] == "fail"
     assert strict_codex["reason"] == "CODEX_AUTH_MISSING"
-
-
-@pytest.mark.unit
-def test_provider_readiness_existing_file_providers_report_credential_scope(
-    tmp_path: Path,
-) -> None:
-    home = tmp_path / "home"
-    (home / ".claude").mkdir(parents=True)
-    (home / ".claude" / "settings.json").write_text('{"token":"claude_file_secret"}')
-    (home / ".claude.json").write_text('{"oauth":"claude_json_secret"}')
-    (home / ".gemini").mkdir()
-    (home / ".gemini" / "oauth_creds.json").write_text("gemini_file_secret")
-    (home / ".config" / "opencode").mkdir(parents=True)
-    (home / ".config" / "opencode" / "opencode.json").write_text("opencode_file_secret")
-    (home / ".ollama").mkdir()
-    (home / ".ollama" / "id_ed25519").write_text("ollama_file_secret")
-
-    payload = collect_agent_readiness(
-        _settings(tmp_path),
-        environ={"AWF_OPENCODE_OLLAMA_BASE_URL": "http://ollama.local:11434/v1"},
-        run_subprocess=_unexpected_subprocess,
-        http_get=_ollama_ok,
-    )
-
-    for name in ("claude_code", "gemini", "opencode"):
-        provider = payload["providers"][name]
-        assert provider["ok"] is True
-        assert provider["credential_scope"] == "isolated_workspace"
-        assert provider["isolation"] == "per_workspace_copy"
-        assert provider["credential_sources"]
-    serialized = json.dumps(payload, sort_keys=True)
-    for secret in (
-        "claude_file_secret",
-        "claude_json_secret",
-        "gemini_file_secret",
-        "opencode_file_secret",
-        "ollama_file_secret",
-    ):
-        assert secret not in serialized
-
-
-@pytest.mark.unit
-def test_provider_readiness_env_fallbacks_report_security_warnings(
-    tmp_path: Path,
-) -> None:
-    """Environment fallback auth reports static-token security warnings."""
-    env = {
-        "AWF_GITHUB_TOKEN": "ghp_env_fallback_secret",
-        "OPENAI_API_KEY": "sk-proj-codex-fallback-secret",
-        "ANTHROPIC_API_KEY": "sk-ant-env-fallback-secret",
-        "CURSOR_API_KEY": "cursor-env-fallback-secret",
-        "GEMINI_API_KEY": "gemini-env-fallback-secret",
-        "OLLAMA_API_KEY": "ollama-env-fallback-secret",
-        "XAI_API_KEY": "xai-env-fallback-secret",
-        "AWF_OPENCODE_OLLAMA_BASE_URL": "http://ollama.local:11434/v1",
-    }
-
-    payload = collect_agent_readiness(
-        _settings(tmp_path),
-        environ=env,
-        run_subprocess=lambda _args, **_kwargs: _completed(stdout="logged in\n"),
-        http_get=_ollama_ok,
-    )
-
-    for name in ("github", "codex", "claude_code", "cursor", "gemini", "opencode", "grok"):
-        provider = payload["providers"][name]
-        assert provider["ok"] is True
-        assert provider["credential_scope"] == "static_env_token"
-        assert provider["isolation"] == "service_env"
-        assert any(warning["reason"] == "STATIC_TOKEN_FALLBACK" for warning in provider["warnings"])
-    serialized = json.dumps(payload, sort_keys=True)
-    for secret in env.values():
-        assert secret not in serialized
-
-
-@pytest.mark.unit
-def test_provider_readiness_docker_reports_host_daemon_broad_control_warning(
-    tmp_path: Path,
-) -> None:
-    payload = collect_agent_readiness(
-        _settings(tmp_path),
-        environ={},
-        run_subprocess=_unexpected_subprocess,
-    )
-
-    docker = payload["providers"]["docker"]
-    assert docker["ok"] is True
-    assert docker["status"] == "ok"
-    assert docker["reason"] == "DOCKER_HOST_CONFIGURED"
-    assert docker["credential_scope"] == "docker_host_control"
-    assert docker["isolation"] == "host_daemon"
-    assert any(warning["reason"] == "DOCKER_HOST_BROAD_CONTROL" for warning in docker["warnings"])
-
-
-@pytest.mark.unit
-def test_provider_readiness_docker_registry_auth_is_observed_not_read(
-    tmp_path: Path,
-) -> None:
-    home = tmp_path / "home"
-    docker_home = home / ".docker"
-    docker_home.mkdir(parents=True)
-    (docker_home / "config.json").write_text('{"auths":{"ghcr.io":{"auth":"docker_file_secret"}}}')
-    env_auth = '{"auths":{"registry.example":{"auth":"docker_env_secret"}}}'
-
-    payload = collect_agent_readiness(
-        _settings(tmp_path),
-        environ={"DOCKER_AUTH_CONFIG": env_auth},
-        run_subprocess=_unexpected_subprocess,
-    )
-
-    docker = payload["providers"]["docker"]
-    source_signals = {source["signal"] for source in docker["credential_sources"]}
-    assert "DOCKER_AUTH_CONFIG" in source_signals
-    assert "~/.docker/config.json" in source_signals
-    serialized = json.dumps(payload, sort_keys=True)
-    assert "docker_file_secret" not in serialized
-    assert "docker_env_secret" not in serialized
-
-
-@pytest.mark.unit
-def test_provider_readiness_docker_reports_missing_auth_without_host_signal(
-    tmp_path: Path,
-) -> None:
-    home = tmp_path / "home"
-    result = provider_readiness._check_docker_provider(
-        replace(_settings(tmp_path), docker_host=""),
-        environ={},
-        host_home=home,
-        strict=True,
-        secrets=frozenset(),
-    )
-
-    assert result["status"] == "fail"
-    assert result["reason"] == "DOCKER_AUTH_NOT_OBSERVED"
-    assert result["credential_scope"] == "not_observed"
-    assert result["isolation"] == "none"
-
-
-@pytest.mark.unit
-def test_provider_readiness_docker_without_host_or_registry_warns(tmp_path: Path) -> None:
-    payload = collect_agent_readiness(
-        _settings(tmp_path, docker_host=""),
-        environ={},
-        run_subprocess=_unexpected_subprocess,
-    )
-
-    docker = payload["providers"]["docker"]
-    assert docker["ok"] is False
-    assert docker["status"] == "warn"
-    assert docker["reason"] == "DOCKER_AUTH_NOT_OBSERVED"
-    assert docker["credential_scope"] == "not_observed"
-    assert docker["isolation"] == "none"
-
-
-@pytest.mark.unit
-def test_provider_readiness_docker_config_path_is_reported_without_reading_secret(
-    tmp_path: Path,
-) -> None:
-    docker_config = tmp_path / "docker-config"
-    docker_config.mkdir()
-    (docker_config / "config.json").write_text(
-        '{"auths":{"registry.example":{"auth":"docker_config_secret"}}}'
-    )
-
-    result = provider_readiness._check_docker_provider(
-        replace(_settings(tmp_path), docker_host=""),
-        environ={"DOCKER_CONFIG": str(docker_config)},
-        host_home=tmp_path / "home",
-        strict=False,
-        secrets=frozenset({"docker_config_secret"}),
-    )
-
-    assert result["status"] == "ok"
-    assert result["reason"] == "DOCKER_REGISTRY_AUTH_PRESENT"
-    assert result["credential_scope"] == "read_only_host_path"
-    assert result["isolation"] == "read_only_bind"
-    assert "docker_config_secret" not in json.dumps(result, sort_keys=True)
-
-
-@pytest.mark.unit
-def test_provider_readiness_docker_config_path_reports_registry_auth(
-    tmp_path: Path,
-) -> None:
-    docker_config = tmp_path / "docker-config"
-    docker_config.mkdir()
-    (docker_config / "config.json").write_text(
-        '{"auths":{"registry.example":{"auth":"docker_config_secret"}}}'
-    )
-
-    payload = collect_agent_readiness(
-        _settings(tmp_path, docker_host=""),
-        environ={"DOCKER_CONFIG": str(docker_config)},
-        run_subprocess=_unexpected_subprocess,
-    )
-
-    docker = payload["providers"]["docker"]
-    assert docker["ok"] is True
-    assert docker["reason"] == "DOCKER_REGISTRY_AUTH_PRESENT"
-    assert docker["credential_scope"] == "read_only_host_path"
-    assert docker["isolation"] == "read_only_bind"
-    assert docker["credential_sources"] == [
-        {
-            "type": "path",
-            "signal": "DOCKER_CONFIG/config.json",
-            "credential_scope": "read_only_host_path",
-            "isolation": "read_only_bind",
-        }
-    ]
-    assert "docker_config_secret" not in json.dumps(payload, sort_keys=True)
-
-
-@pytest.mark.unit
-def test_provider_readiness_docker_config_env_does_not_fall_back_to_home(
-    tmp_path: Path,
-) -> None:
-    docker_config = tmp_path / "missing-docker-config"
-    home_docker = tmp_path / "home" / ".docker"
-    home_docker.mkdir(parents=True)
-    (home_docker / "config.json").write_text("home_docker_secret")
-
-    result = provider_readiness._check_docker_provider(
-        replace(_settings(tmp_path), docker_host=""),
-        environ={"DOCKER_CONFIG": str(docker_config)},
-        host_home=tmp_path / "home",
-        strict=False,
-        secrets=frozenset({"home_docker_secret"}),
-    )
-
-    assert result["status"] == "warn"
-    assert result["reason"] == "DOCKER_AUTH_NOT_OBSERVED"
-    assert "home_docker_secret" not in json.dumps(result, sort_keys=True)
-
-
-@pytest.mark.unit
-def test_provider_readiness_explicit_missing_docker_config_does_not_fallback(
-    tmp_path: Path,
-) -> None:
-    home = tmp_path / "home"
-    docker_home = home / ".docker"
-    docker_home.mkdir(parents=True)
-    (docker_home / "config.json").write_text(
-        '{"auths":{"registry.example":{"auth":"docker_home_secret"}}}'
-    )
-
-    payload = collect_agent_readiness(
-        _settings(tmp_path, host_home=str(home), docker_host=""),
-        environ={"DOCKER_CONFIG": str(tmp_path / "missing-docker-config")},
-        run_subprocess=_unexpected_subprocess,
-    )
-
-    docker = payload["providers"]["docker"]
-    assert docker["ok"] is False
-    assert docker["reason"] == "DOCKER_AUTH_NOT_OBSERVED"
-    assert docker["credential_sources"] == []
-    assert "docker_home_secret" not in json.dumps(payload, sort_keys=True)
